Move Sudoku to own folder
